third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-stretch-alias-expected.txt

This is a testharness.js-based test.
[FAIL] font-stretch is a legacy name alias for font-width
  assert_equals: font-stretch is an alias for font-width expected "inherit" but got ""
Harness: the test ran to completion.
